Clase 9 Java
modifico código calculadora, convierto las variables en Double, para poder realizar operaciones mas precisas

+++ Java/Leccion09/CalculadoraUTN/src/CalculadoraUTN.java
@@ -38,10 +38,10 @@ public class CalculadoraUTN {
     }//Fin metodo mostrarMenu
     private static void ejecutarOperacion(int operacion, Scanner entrada){
             System.out.print("Digite el valor para el operando 1: ");
-            var operando1 = Integer.parseInt(entrada.nextLine());
+            var operando1 = Double.parseDouble(entrada.nextLine());
             System.out.print("Digite el valor para el operando 2: ");
-            var operando2 = Integer.parseInt(entrada.nextLine());
-            int resultado;
+            var operando2 = Double.parseDouble(entrada.nextLine());
+            Double resultado;
             switch (operacion) {
                     case 1 -> { //Suma
                             resultado = operando1 + operando2;
